resolves #30 resolves #32 주석 및 코드 정리, 최종 배포 설정(CI, CD) (#33)
* Spring init

* #1 Jwt and Security config

* #1 User Entity, Dto, Service, Repository

* #1 User Controller init

* java version update

* #1 Signup, Signin, Response Dto

* #1 Update User Entity, Repository

* #1 Signup, Signin Controller

* #1 application.yaml config h2 database

* #1 fix CustomUserdetails

* #3 Change project folder name

* #3 Config CORS mapping

* #3 Create react app

* fix files to commit

* delete error folder

* #3 Set SigninResponse return with username

* #3 frontend layouts

* #3 frontend signup, signin design and link to backend

* #3 회원명 중복 검사 로직

* #3 주석 수정 및 약간의 오류 수정

* #3 회원명 중복 검사 api 연동

* #3 내비게이션 로고 크기 및 클릭 범위 수정

* #5 Posts Entity

* #5 Posts Controller and Dto

* #5 Posts Service and Repository

* #7 Posts CRUD Page

* #7 주석 수정

* #7 Update Posts Dto

* #7 검색 및 페이지네이션 기능 구현

* #7 검색과 페이지네이션과 Dto 업데이트 및 수정

* #7 Floara WYSIWYG 기본 설정 및 약간의 오류 수정

* #7 Post 전체 가져오기 및 검색 기능 완성

* #7 Floara WYSIWYG을 사용한 Post 생성 기능 완성

* 리베이스 진행 용 커밋

* resolves #5 Create post backend function (#6)

* Spring init

* #1 Jwt and Security config

* #1 User Entity, Dto, Service, Repository

* #1 User Controller init

* java version update

* #1 Signup, Signin, Response Dto

* #1 Update User Entity, Repository

* #1 Signup, Signin Controller

* #1 application.yaml config h2 database

* #1 fix CustomUserdetails

* #3 Change project folder name

* #3 Config CORS mapping

* #3 Create react app

* fix files to commit

* delete error folder

* #3 Set SigninResponse return with username

* #3 frontend layouts

* #3 frontend signup, signin design and link to backend

* #3 회원명 중복 검사 로직

* #3 주석 수정 및 약간의 오류 수정

* #3 회원명 중복 검사 api 연동

* #3 내비게이션 로고 크기 및 클릭 범위 수정

---------


* Revert "resolves #5 Create post backend function (#6)" (#8)

This reverts commit 389d164952af0a267635cef4181353c048d57ce8.

* 병합용 커밋

* #7 Json Web Token 전송 방식을 Cookie로 변경(보안 향상)

* #7 database ddl-auto 설정 변경

* #7 Post 관리 페이지 디자인 및 기능 추가, 수정

* #7 Post 삭제 및 좋아요 기능 구현

* #7 주석 삭제 및 추가

* #11 모바일 화면에서도 아이콘 보이게 수정

* #11 Comment, CommentLikes Entity

* #11 파일 이름 변경 및 위치 변경

* #11 Comment Controller, Dto

* #11 Comment Service, Repository

* #13 Comment CRUD page and link with backend function

* #14 Create myPage function + 내 게시판, 본 게시글 저장, 수정 내역 저장

* #16 Create my page and etc design link with backend function

* 병합용 커밋

* #18 My page, My Favorite, My Actives

* 병합용 커밋

* #18 Create Favorite and My Actives Function

* #18 Create Favorite and My Actives Pages and delete alert

* #19 Alert Function 일부 및 질문 게시판에서 댓글 채택 기능

* #19 Alert Page 및 질문 게시글에서의 채택 버튼 표시 및 이미 채택된 댓글 다르게 표시

* #19 공통된 색상 및 함수를 독립된 하나의 파일로 만들어 관리

* #19 Posts에 maxUserNumber(최대 인원수), currentUserNumber(현재 인원수) 추가

* #19 모임 신청 로직

* #19 알림 Entity, Dto, Controller, Service, Repository

* #19 신청 알림 승인 및 거절 로직

* #19 알림 읽음 상태 로직 비활성화(오류 발생: 추가 후에 바로 삭제하면 안됨)

* #19 알림을 읽지 않았을 때 녹색 표시 비활성화(오류)

* #19 모임 신청 상태에 따라 모임 신청란 다르게 표시

* #19 회원가입 이메일 검증 Page

* #19 읽은 알림 구분을 기능 및 UserViewed -> PostsViewed로 이름 변경

* #19 useLocation을 이용해서 매 랜더링마다 새로운 알림이 있는지 검사 및 표시, 읽지 않은 알림을 NEW_COLOR로 표시

* #19 UserViewedRepository.java -> PostsViewedRepository.java 이름 변경

* #22 Email Validate page

* #22 이메일에 코드 전송을 할 때 이메일 중복 검사 실행

* #22 git ignore

* #22 이메일 서버(Gmail SMTP) 및 Redis 연결 정보 설정

* #22 인증 코드 생성 클래스

* #22 인증코드 전송 및 검사 서비스

* #22 약간의 오류 수정

* #22 이메일 인증 컨트롤러 및 이메일 의존성 추가

* #22 이메일 인증 api 연결

* #22 이메일 인증 후 권한 TEMP -> 비밀번호 재설정 로직에 TEMP 권한을 가진 사용자만 접근할 수 있도록 설정

* #22 비밀번호 재설정 페이지, axios 최대 대기 가능 시간 증가(이메일 보내는게 오래걸림), 디자인 수정

* #22 비밀번호 재설정 페이지

* #24 Implement Web Socket

* #24 백엔드 웹소켓 설정 (기능 api 미구현)

* #24 백엔드 웹소켓 설정 및 기능 일부 생성 (메모리 -> 데이터베이스)

* #24 프론트엔드 리스트 필드 크기 조절 및 폰트 notoSansKr로 변경, 커뮤니티 리스트 페이지 및 커뮤니티 대화 페이지 생성

* #24 약간의 오류 수정 및 정리

* #24 Real-time community function, 게시글 삭제 오류 있음

* #24 Real-time community Page

* #26 댓글 삭제 -> 알림 삭제, 채택 댓글 오류 수정

* #26 Implement S3

* #26 Implement aws auth

* #26 S3 Service, Controller, Dto, Aws config

* #26 api 요청 추가

* 병합용 커밋

* #26 S3 기능 완성

* #28 관리자 페이지, 디자인 및 약간의 오류 수정

* #28 관리자 기능 및 프론트엔드로 현재 사용자가 관리자임을 알려주기 위한 기능 추가

* #28 알림 조회 오류 수정

* #30 프론트엔드 utilities/.. 제외한 파일 주석 및 비밀번호 입력시 소문자 변환 설정

* #30 프론트엔드 주석 정리 완료

* #30 백엔드 컨트롤러 부분 주석 추가 및 수정

* #30 백엔드 엔티티 부분 주석 추가 및 수정

* #30 백엔드 시큐리티 부분 주석 추가 및 수정

* #30 백엔드 서비스 부분 주석 추가 및 수정

* #30 백엔드 Web 부분 주석 추가 및 수정 - 주석 작업 완료

* #30 채팅방에 초대된 사용자 제거할 때 발생하던 오류 수정 - 컬렉션 순환중 다중 수정 오류

* #30 비밀번호 한글 금지, 최소 8자 이상 기능 추가

* #32 배포 설정

* #32 백엔드 elastic beanstalk 이름 수정

* #32 주석 추가

* #32 elastic beanstalk 오류 해결 및 최종 수정

---------

+++ backend/src/main/java/com/example/backend/service/PostsService.java
@@ -10,8 +10,8 @@ import com.example.backend.dto.posts.update.PostsUpdateRequest;
 import com.example.backend.entity.*;
 import com.example.backend.entity.utilities.PostsSubject;
 import com.example.backend.repository.*;
-import com.example.backend.service.utilities.PostLikesSpec;
-import com.example.backend.service.utilities.PostSearchSpec;
+import com.example.backend.service.searchSpec.PostLikesSpec;
+import com.example.backend.service.searchSpec.PostSearchSpec;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.data.domain.Page;
@@ -25,6 +25,7 @@ import java.util.List;
 
 import static com.example.backend.entity.utilities.AlertSubject.*;
 import static com.example.backend.entity.utilities.PostsSubject.*;
+import static com.example.backend.entity.utilities.Role.ROLE_ADMIN;
 
 @Slf4j
 @RequiredArgsConstructor
@@ -113,6 +114,16 @@ public class PostsService {
                 .build());
     }
 
+    /**
+     * 특정 사용자의 즐겨찾기 게시글 목록을 검색 조건과 페이징 조건에 따라 조회합니다.
+     *
+     * @param user 현재 로그인된 사용자 (좋아요 여부 확인 및 대상 사용자 지정)
+     * @param pageable 페이징 정보
+     * @param searchField 검색 필드
+     * @param searchTerm 검색어
+     * @param tab 주제별 탭 필터
+     * @return 검색 및 페이징 처리된 해당 사용자의 게시글 목록 DTO
+     */
     public Page<PostsIndexResponse> indexFavoriteByUser(User user, Pageable pageable, String searchField, String searchTerm, Integer tab) {
         Specification<PostsLikes> spec = PostLikesSpec.search(user, searchField, searchTerm, tab);
 
@@ -342,7 +353,7 @@ public class PostsService {
         Long deletedId = target.getId();
         // 작성자 권한 검증
         // 오류 발생 -> userDetails에서 가져온 user는 <비영속>, 데이터베이스에서 조회된 user는 <영속> 상태임
-        // if(!user.equals(target.getUser())) throw new IllegalAccessException("다른 사용자의 글을 삭제할 수 없습니다.");
+         if(!user.getId().equals(target.getUser().getId()) || !user.getAuthority().equals(ROLE_ADMIN)) throw new IllegalAccessException("다른 사용자의 글을 삭제할 수 없습니다.");
 
         repository.delete(target);
